new augmentation method
Another attempt at issue #5.

idea: instead of only adding random negatives we now only add
positives to positives and negatives to negatives.  The goal is to not
override context cues that may be present.

I probably won't get around to evaluating
this until next week, so if someone wants to try running it and see if
it does better than the previous approach that would be great!